+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 🔧 Internal changes
 
+- Refactor GraphDropdown component from being a child of Graph to being a child of NavBar
+
 ## [0.7.2] - 2025-12-10
 
 ### ✨ New features/enhancements